Fix NULL pointer deference in tsem_task_kill hook.
Testing with the Quixote BusyBox implementation resulted in a
NULL pointer dereference when the SI_FROMKERNEL test was
conducted on the kernel_siginfo pointer.

This was secondary to the fact that a kernel_siginfo pointer can
assume the SEND_SIG_NOINFO value that is the null pointer value.
Previous testing had not run into this scenario.

Make the SI_FROMKERNEL(info) check be contingent on the fact that
the info pointer is not the SEND_SIG_NOINFO value.