Rename task trapping functions.
Feedback from review of the second patch series included concern
over the following two functions:

return_trapped_task

return_trapped_inode

The objection was that the function names didn't make any sense.

I had noted that the functions had been named the way they were
since they were declarative names indicating what the functions
did, ie. return the return status of the task when it was trapped
secondary to being untrusted and when an untrusted task attempted
to access an inode.

Since the functions are always called as an argument to the
'return' command the return_ prefix is redundant and can be
removed.