Fix handling of tsem_inode_setxattr.
The previous implementation of the tsem_inode_setxattr handler
was not properly dealing with externally modeled namespaces since
the attribute value was not being Base64 encoded.

Transfer the Base64 encoding to the get_inode_setxattr() function
in the event.c file.  An extra character pointer was added to the
output structure of the tsem_inode_getxattr_args structure to
hold the Base64 encoded value of the attribute value.  The
mapping function for the security handler frees the memory that
was allocated for the raw attribute value.

The size value was also added to the output member of the
argument structure since that value is needed in order to
properly extend the attribute value into the generation of the
coefficient.